# pkg/docs/swaggers/auth.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "auth.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "AuthAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/auth/config": {
      "get": {
        "operationId": "AuthAPI_GetConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/artel_authGetConfigResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "AuthAPI"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthAPI_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/artel_authLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/artel_authLoginRequest"
            }
          }
        ],
        "tags": [
          "AuthAPI"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthAPI_Logout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/artel_authLogoutResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/artel_authLogoutRequest"
            }
          }
        ],
        "tags": [
          "AuthAPI"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthAPI_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/artel_authRegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/artel_authRegisterRequest"
            }
          }
        ],
        "tags": [
          "AuthAPI"
        ]
      }
    }
  },
  "definitions": {
    "artel_authGetConfigResponse": {
      "type": "object",
      "properties": {
        "telegramClientId": {
          "type": "string"
        }
      }
    },
    "artel_authLoginRequest": {
      "type": "object",
      "properties": {
        "password": {
          "$ref": "#/definitions/artel_authPasswordCredentials"
        },
        "telegram": {
          "$ref": "#/definitions/artel_authTelegramCredentials"
        }
      }
    },
    "artel_authLoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "expiresAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "artel_authLogoutRequest": {
      "type": "object"
    },
    "artel_authLogoutResponse": {
      "type": "object"
    },
    "artel_authPasswordCredentials": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "artel_authRegisterRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "artel_authRegisterResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "artel_authTelegramCredentials": {
      "type": "object",
      "properties": {
        "idToken": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}